Make bookie watcher an interface
So that it can be mocked out easily for testing.
This patch also contains a simple mock.
